[refactor] modify code structure
